+++ modules/utils/src/filesystem/file.cpp
@@ -2,8 +2,6 @@
 
 #include <fstream>
 
-#include <absl/random/random.h>
-#include <absl/time/clock.h>
 #include <fmt/core.h>
 
 #include "hephaestus/utils/exception.h"
@@ -11,8 +9,8 @@
 namespace heph::utils::filesystem {
 auto readFile(const std::filesystem::path& path) -> std::string {
   std::ifstream infile{ path };
-  throwExceptionIf<InvalidDataException>(
-      !infile, fmt::format("Could not open file {} to read content", path.string()));
+  throwExceptionIf<InvalidDataException>(!infile,
+                                         fmt::format("Could not open file {} to read", path.string()));
 
   std::string text;
   text.reserve(std::filesystem::file_size(path));
@@ -23,8 +21,8 @@ auto readFile(const std::filesystem::path& path) -> std::string {
 
 auto readBinaryFile(const std::filesystem::path& path) -> std::vector<std::byte> {
   std::ifstream infile{ path, std::ios::binary };
-  throwExceptionIf<InvalidDataException>(
-      !infile, fmt::format("Could not open file {} to read content", path.string()));
+  throwExceptionIf<InvalidDataException>(!infile,
+                                         fmt::format("Could not open file {} to read", path.string()));
 
   const auto file_size = std::filesystem::file_size(path);
   std::vector<std::byte> buffer(file_size);
@@ -49,61 +47,6 @@ void writeBufferToFile(const std::filesystem::path& path, std::span<const std::b
 
   // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
   outfile.write(reinterpret_cast<const char*>(content.data()), static_cast<std::streamsize>(content.size()));
-}
-
-// -------------------------------------------------------------------------------------------------
-// ScopedPath implementation
-// -------------------------------------------------------------------------------------------------
-
-ScopedPath::ScopedPath(std::filesystem::path path) : path_(std::move(path)), path_str_(path_.string()) {
-}
-
-ScopedPath::~ScopedPath() {
-  std::error_code e;
-  std::filesystem::remove_all(path_, e);
-  if (e) {
-    fmt::print(stderr, "failed to remove test temporary directory {}, with error: {}\n", path_.string(),
-               e.message());
-  }
-}
-
-auto ScopedPath::createFile() -> ScopedPath {
-  const auto global_temp_dir = std::filesystem::temp_directory_path();
-  auto file = global_temp_dir / randomString();
-  while (std::filesystem::exists(file)) {
-    file = global_temp_dir / randomString();
-  }
-
-  return ScopedPath{ file };
-}
-
-auto ScopedPath::createDir() -> ScopedPath {
-  const auto global_temp_dir = std::filesystem::temp_directory_path();
-  auto temp_dir = global_temp_dir / randomString();
-  while (std::filesystem::exists(temp_dir)) {
-    temp_dir = global_temp_dir / randomString();
-  }
-
-  std::filesystem::create_directory(temp_dir);
-  return ScopedPath{ std::move(temp_dir) };
-}
-
-ScopedPath::operator std::filesystem::path() const {
-  return path_;
-}
-
-ScopedPath::operator std::string() const {
-  return path_;
-}
-
-ScopedPath::operator std::string_view() const {
-  return { path_str_ };
-}
-
-auto ScopedPath::randomString() -> std::string {
-  absl::BitGen bitgen;
-  return fmt::format("{}_{}", absl::ToUnixMillis(absl::Now()),
-                     bitgen() % 1000);  // NOLINT(cppcoreguidelines-avoid-magic-numbers)
 }
 
 }  // namespace heph::utils::filesystem
